Add the Top Vote Path method, which is monotone and resists manipulation.
This combination of monotonicity and electing from the resistant set
can to my knowledge not be found anywhere else (as of this commit). TVP
thus establishes that the criteria are compatible.

The Python file also contains tests which should help those who'd like
to implement TVP or port it to another language.